+++ epistola-client-kotlin/client/src/main/kotlin/io/epistola/client/EpistolaClient.kt
@@ -1,0 +1,5 @@
+package io.epistola.client
+
+class EpistolaClient {
+    // TODO: Implement client
+}