+++ epistola-client-kotlin/client/src/test/kotlin/io/epistola/client/EpistolaClientTest.kt
@@ -1,0 +1,12 @@
+package io.epistola.client
+
+import kotlin.test.Test
+import kotlin.test.assertNotNull
+
+class EpistolaClientTest {
+    @Test
+    fun `client can be instantiated`() {
+        val client = EpistolaClient()
+        assertNotNull(client)
+    }
+}